refactor: split MainView code-behind into focused partials
MainView.axaml.cs had grown to ~800 lines mixing five unrelated
responsibilities. Split by concern into partial-class files and extract
the palette drag as a standalone collaborator. No behavior change.

- MainView.axaml.cs: core wiring, keyboard shortcuts, lifecycle, toasts
- MainView.Commands.cs: thin menu/toolbar click handlers
- MainView.FileCommands.cs: new/open/close/save/save-as/screenshot + dialog
- MainView.Preview.cs: animation-preview timer
- PaletteDragController.cs: palette->canvas drag, extracted as its own class

Retarget the two source-text assertion tests to the new file layout.
