+++ src/CtrDxEditor.Shared/Views/MainView.Commands.cs
@@ -1,0 +1,120 @@
+using Avalonia.Controls;
+using Avalonia.Interactivity;
+
+using CtrDxEditor.Core.Document;
+using CtrDxEditor.Rendering;
+using CtrDxEditor.ViewModels;
+
+namespace CtrDxEditor.Views
+{
+    // Thin menu/toolbar click handlers that delegate straight to the view model or canvas. Keyboard
+    // equivalents live in the KeyDown handler wired up in the constructor.
+    public partial class MainView
+    {
+        private void Undo_Click(object? sender, RoutedEventArgs e)
+        {
+            if (DataContext is EditorViewModel vm)
+            {
+                vm.Undo();
+            }
+        }
+
+        private void Redo_Click(object? sender, RoutedEventArgs e)
+        {
+            if (DataContext is EditorViewModel vm)
+            {
+                vm.Redo();
+            }
+        }
+
+        private void Delete_Click(object? sender, RoutedEventArgs e)
+        {
+            if (DataContext is EditorViewModel vm)
+            {
+                vm.DeleteSelected();
+                this.FindControl<LevelCanvas>("Canvas")!.InvalidateVisual();
+            }
+        }
+
+        private void SnapToggle_Click(object? sender, RoutedEventArgs e)
+        {
+            if (DataContext is EditorViewModel vm)
+            {
+                vm.SnapEnabled = !vm.SnapEnabled;
+            }
+        }
+
+        private void ShowHitboxesToggle_Click(object? sender, RoutedEventArgs e)
+        {
+            if (DataContext is EditorViewModel vm)
+            {
+                vm.ShowHitboxes = !vm.ShowHitboxes;
+            }
+        }
+
+        private void ShowMobileHitboxesToggle_Click(object? sender, RoutedEventArgs e)
+        {
+            if (DataContext is EditorViewModel vm)
+            {
+                vm.ShowMobileHitboxes = !vm.ShowMobileHitboxes;
+            }
+        }
+
+        private void ShowForceFieldsToggle_Click(object? sender, RoutedEventArgs e)
+        {
+            if (DataContext is EditorViewModel vm)
+            {
+                vm.ShowForceFields = !vm.ShowForceFields;
+            }
+        }
+
+        private void ShowMovementPathsToggle_Click(object? sender, RoutedEventArgs e)
+        {
+            if (DataContext is EditorViewModel vm)
+            {
+                vm.ShowMovementPaths = !vm.ShowMovementPaths;
+            }
+        }
+
+        private void AnimationPreviewToggle_Click(object? sender, RoutedEventArgs e)
+        {
+            if (DataContext is EditorViewModel vm)
+            {
+                vm.ToggleAnimationPreviewAll();
+            }
+        }
+
+        private void ObjectAnimationPreview_Click(object? sender, RoutedEventArgs e)
+        {
+            if (DataContext is EditorViewModel vm && sender is Button { Tag: LevelObject obj })
+            {
+                vm.ToggleAnimationPreviewObject(obj);
+                e.Handled = true;
+            }
+        }
+
+        private void ObjectLock_Click(object? sender, RoutedEventArgs e)
+        {
+            if (DataContext is EditorViewModel vm && sender is Button { Tag: LevelObject obj })
+            {
+                vm.ToggleLock(obj);
+                e.Handled = true;
+            }
+        }
+
+        private void ZoomIn_Click(object? sender, RoutedEventArgs e)
+        {
+            this.FindControl<LevelCanvas>("Canvas")!.ZoomBy(1.2);
+        }
+
+        private void ZoomOut_Click(object? sender, RoutedEventArgs e)
+        {
+            this.FindControl<LevelCanvas>("Canvas")!.ZoomBy(1 / 1.2);
+        }
+
+        private void ZoomFit_Click(object? sender, RoutedEventArgs e)
+        {
+            this.FindControl<LevelCanvas>("Canvas")!.FitToView();
+        }
+    }
+}
